Create Simple push button timer
This code is an Arduino sketch for a timer that turns on a relay for a certain duration. When the sketch starts, it reads the timer duration from the EEPROM and sets the timerDurationInHours variable accordingly. In the setup function, the buttonPin, relayPin, and ledPin are configured as inputs and outputs, respectively.

In the loop function, the state of the buttonPin is read and stored in the buttonState variable. If the button was held in for more than 5 seconds at startup, the programmingMode flag is set to true and the selected duration is set to 1. The LED is then turned on and off twice to indicate that the timer is in programming mode.

If the programmingMode flag is true, the code enters a loop where it waits for the button to be pressed and held. While the button is held, the selected duration is incremented and the LED is turned on and off a number of times equal to the current value of the selectedDuration variable. This allows the user to choose the desired duration by holding the button and counting the number of LED blinks.

Once the button is released, the programmingMode flag is set to false and the selected duration is saved to the EEPROM. The code then enters another loop where it waits for the button to be pressed again. When the button is pressed, the timer starts and the relay is turned on. The code then waits for the duration specified in the timerDurationInHours variable to elapse. Once the duration has elapsed, the relay is turned off and the timer is reset.

The previousState variable is used to keep track of the button state in the previous iteration of the loop function. This is used to detect changes in the button state, such as when it is pressed or released.